requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
boto3>=1.34.0
openai>=1.30.0
pydantic>=2.0.0
python-dotenv>=1.0.0
